// BackEnd/src/interfaces/modelTypes/customer.d.ts
const enum customerColumnNames {
  CustId = 'CustId',
  CustName = 'CustName',
  CustMobileNo = 'CustMobileNo',
  CustSociety = 'CustSociety',
  CustBlockNo = 'CustBlockNo',
  CustFlatNo = 'CustFlatNo',
  CustAddress = 'CustAddress',
  CustPincode = 'CustPincode',
  CustCity = 'CustCity',
  CustMailAddress = 'CustMailAddress',
  paymode1 = 'paymode1',
  accname1 = 'accname1',
  paymode2 = 'paymode2',
  accname2 = 'accname2',
}

export { customerColumnNames };